backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pandas==2.2.0
pyarrow==15.0.0
python-dotenv==1.0.0
python-multipart==0.0.9
alembic==1.13.1
ldap3==2.9.1
